refactor(analytics-core): restructure
