+++ packages/analytics-core/src/core-client.ts
@@ -1,8 +1,8 @@
-import { Plugin } from './plugins/plugin';
+import { Plugin } from './types/plugin';
 import { IConfig } from './config';
-import { BaseEvent, EventOptions } from './event/base-event';
+import { BaseEvent, EventOptions } from './types/event/base-event';
 import { Result } from './types/result';
-import { Event } from './event/event';
+import { Event } from './types/event/event';
 import { IIdentify } from './identify';
 import { IRevenue } from './revenue';
 import { CLIENT_NOT_INITIALIZED, OPT_OUT_MESSAGE } from './types/messages';
