+++ packages/analytics-core/src/timeline.ts
@@ -1,8 +1,8 @@
-import { BeforePlugin, DestinationPlugin, EnrichmentPlugin, Plugin } from './plugins/plugin';
+import { BeforePlugin, DestinationPlugin, EnrichmentPlugin, Plugin } from './types/plugin';
 import { CoreClient } from './core-client';
 import { IConfig } from './config';
 import { EventCallback } from './types/event-callback';
-import { Event } from './event/event';
+import { Event } from './types/event/event';
 import { Result } from './types/result';
 import { buildResult } from './utils/result-builder';
 import { UUID } from './utils/uuid';
