+++ packages/analytics-core/src/transports/fetch.ts
@@ -1,5 +1,5 @@
 import { BaseTransport } from './base';
-import { Transport } from './transport';
+import { Transport } from '../types/transport';
 import { Payload } from '../types/payload';
 import { Response } from '../types/response';
