+++ packages/analytics-core/src/types/event/base-event.ts
