+++ packages/analytics-core/src/types/event/ingestion-metadata.ts
