+++ packages/analytics-core/src/types/event/plan.ts
