+++ packages/analytics-core/src/types/result.ts
@@ -1,4 +1,4 @@
-import { Event } from '../event/event';
+import { Event } from './event/event';
 
 export interface Result {
   event: Event;
