+++ packages/analytics-core/src/types/storage.ts
